feat: support chunked prefill for CUDA MLA
Enable ordinary BF16 MLA through model-specific configuration checks while
preserving native TP and block constraints. Keep the required FP8 blockwise
projection path and large-row SiLU offset fix.

Preserve chunk progress and cache ownership across prefix reuse and request
cancellation, with configuration, KV, projection and engine regression tests.

+++ rtp_llm/models_py/triton_kernels/common/activation.py
@@ -18,7 +18,7 @@ def _silu_and_mul_kernel(
     # Meta-parameter for tuning
     BLOCK_SIZE_N: tl.constexpr,
 ):
-    pid_b = tl.program_id(axis=0)  # Batch dimension
+    pid_b = tl.program_id(axis=0).to(tl.int64)  # Row offsets may exceed int32.
     pid_n_block = tl.program_id(axis=1)  # N-dimension block
 
     input_row_start_ptr = input_ptr + pid_b * input_row_stride
